requirements.txt
pandas
numpy
scikit-learn
xgboost
optuna
shap
matplotlib
fastapi
uvicorn
kafka-python
pydantic
python-dotenv
kaggle
